python/agileetc/pyproject.toml
[tool.poetry]
name = "agileetc"
version = "0.0.7"
license = "Apache-2.0"
readme = "README.md"
description = "Future Agile CICD tooling template"
authors = ["agileturret <Paul.Gilligan@agilesolutions.co.uk>"]
homepage = "https://github.com/Agile-Solutions-GB-Ltd/automatic-eureka/tree/main/python/agileetc"
repository = "https://github.com/Agile-Solutions-GB-Ltd/automatic-eureka"

[tool.poetry.scripts]
agileetc = "agileetc.cli:cli"

[tool.pytest.ini_options]
pythonpath = "agileetc"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1"
PyYAML = "^5.4.1"
pytest = "^7.1.3"
prettytable = "^3.3.0"
Fabric3 = "^1.14.post1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0"
bumpversion = "^0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry_bumpversion.file."src/agileetc/__init__.py"]